+++ config/utils/amountInWords.php
@@ -16,7 +16,7 @@ if (!function_exists('numberToWords')) {
         $numArray = str_split(strrev((string)$number), 3);
 
         foreach ($numArray as $key => $value) {
-            $chunk = (int)strrev($value);
+            $chunk = strrev($value);
             if ($chunk != 0) {
                 $chunkWords = array();
